go-gen: introduce script for better go recipes
using information from proxy.golang.org this tool generates recipes
of the requested tool plus all dependencies.
As by nature the dependency tree are very deep in go information is
cached in a description file, which largely speedsup generation.

The script generates recipes for the latest version available (if not
told otherwise by the supplied errata json).
It furthermore uses scancode toolkit to determine the used
licenses in the sources.

Closes #4776

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>